WEP-4 more refactor models and schemas.
