+++ backend/week_eat_planner/api/schemas/recipe.py
@@ -1,5 +1,7 @@
 from uuid import UUID
+
 from pydantic import BaseModel, ConfigDict
+
 
 class OwnerId(BaseModel):
     user_id: UUID
@@ -7,6 +9,7 @@ class OwnerId(BaseModel):
 
 class RecipeBase(BaseModel):
     """Base schema for a recipe, containing common fields."""
+
     name: str
     is_public: bool = False
     ingredients: dict[str, int | float | str]
@@ -14,6 +17,7 @@ class RecipeBase(BaseModel):
 
 class RecipeCreate(RecipeBase):
     """Schema for creating a new recipe. Inherits all fields from RecipeBase."""
+
     pass
 
 
@@ -23,11 +27,13 @@ class RecipeUpdate(RecipeBase):
     Note: In a real-world PATCH scenario, you might want all fields to be optional.
     If so, inherit from a different base or manually define fields as Optional[...].
     """
+
     pass
 
 
 class RecipeRead(RecipeBase, OwnerId):
     """Schema for reading a recipe, including the database ID and user ID."""
+
     id: UUID
 
     model_config = ConfigDict(from_attributes=True)
@@ -35,6 +41,7 @@ class RecipeRead(RecipeBase, OwnerId):
 
 class RecipeReadMinimal(BaseModel):
     """A minimal schema for recipe previews, showing only the ID and name."""
+
     id: UUID
     name: str
