+++ backend/week_eat_planner/api/schemas/week.py
@@ -1,10 +1,10 @@
+from typing import TYPE_CHECKING
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict
 
-from week_eat_planner.db.models import DayOfWeek, MealType
-
-from .recipe import RecipeReadMinimal
+if TYPE_CHECKING:
+    from .meal_slot import MealSlotRead
 
 
 class WeekBase(BaseModel):
@@ -37,26 +37,9 @@ class WeekReadMinimal(WeekBase, OwnerId):
     model_config = ConfigDict(from_attributes=True)
 
 
-class MealSlotRead(BaseModel):
-    """Schema for a meal slot representation."""
-
-    id: UUID
-    day_of_week: DayOfWeek
-    meal_type: MealType
-    recipe: RecipeReadMinimal | None = None
-
-    model_config = ConfigDict(from_attributes=True)
-
-
-class MealSlotAssign(BaseModel):
-    """Schema for assigning a recipe to a meal slot."""
-    slot_id: UUID
-    recipe_id: UUID
-
-
 class WeekRead(WeekReadMinimal):
     """Schema for a detailed representation of a week, including meal slots."""
 
-    meal_slots: list[MealSlotRead]
+    meal_slots: list['MealSlotRead']
 
     model_config = ConfigDict(from_attributes=True)
