+++ backend/week_eat_planner/db/models/recipe.py
@@ -1,3 +1,4 @@
+from typing import TYPE_CHECKING
 from uuid import UUID
 
 from sqlalchemy import ForeignKey
@@ -5,6 +6,9 @@ from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from week_eat_planner.db.base import Base
+
+if TYPE_CHECKING:
+    from .user import User
 
 
 class Recipe(Base):
